+++ src/crown.py
@@ -14,7 +14,7 @@ from auto_LiRPA import BoundedModule, BoundedTensor
 from auto_LiRPA.perturbations import PerturbationLpNorm
 
 device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
-torch.set_default_tensor_type(torch.cuda.DoubleTensor)
+torch.set_default_tensor_type(torch.cuda.FloatTensor)
 
 batch_size_per_iteration = 100000
 
@@ -85,8 +85,8 @@ class CrownImplicitFunction(implicit_function.ImplicitFunction):
 
 
     def classify_box(self, params, box_lower, box_upper, offset=0.):
-        ptb = PerturbationLpNorm(x_L=box_lower.double(), x_U=box_upper.double())
-        bounded_x = BoundedTensor(box_lower.double(), ptb)
+        ptb = PerturbationLpNorm(x_L=box_lower.float(), x_U=box_upper.float())
+        bounded_x = BoundedTensor(box_lower.float(), ptb)
         may_lower, may_upper = self.bounded_func.compute_bounds(x=(bounded_x,), method=self.crown_mode, bound_upper=False)
         bound_dict = self.bounded_func.save_intermediate()
         ptb = PerturbationLpNorm(x_L=box_lower.float(), x_U=box_upper.float())
